add metadata for create and edit flows, fix table refetch after creating a new dataset
