+++ app/src/components/dataset/styles.tsx
@@ -1,0 +1,7 @@
+import { css } from "@emotion/react";
+
+export const metadataFieldWrapperCSS = css`
+  .ac-field__field {
+    width: 100%;
+  }
+`;
